new/pytest-kube-provider/pyproject.toml
[tool.poetry]
name = "pytest-kube-provider"
version = "0.1.0"
description = "A simple plugin to use with pytest]: A plugin to provide different types and configs of Kubernetes clusters that can be used for testing."
authors = ["Łukasz Piątkowski <lukasz@giantswarm.io>", "Tobias Bradtke <tobias@giantswarm.io"]
license="Apache-2.0"
repository="https://github.com/giantswarm/pytest-kube"
classifiers=[
    'Development Status :: 4 - Beta',
    'Framework :: Pytest',
    'Intended Audience :: Developers',
    'Topic :: Software Development :: Testing',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: Implementation :: CPython',
    'Operating System :: OS Independent',
    'License :: OSI Approved :: Apache Software License',
]

[tool.poetry.plugins] # Optional super table

[tool.poetry.plugins."pytest11"]
"kube-provider" = "pytest_kube_provider"

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^5.4.2"
pykube-ng = "^20.5.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
